Split linksystem into multiple files
